% Generated by roxygen2 (4.0.1): do not edit by hand
\name{adjacent2edge}
\alias{adjacent2edge}
\title{Function to convert tree in adjacent list format to edge
representation.}
\usage{
adjacent2edge(tree)
}
\arguments{
\item{The}{graph or tree in adjacent list format.}
}
\description{
Function to convert tree in adjacent list format to edge
representation.
}

