% Generated by roxygen2 (4.0.1): do not edit by hand
\name{computeRatio}
\alias{computeRatio}
\title{Function for computing the yield}
\usage{
computeRatio(numerator, denominator)
}
\arguments{
\item{production}{The value of production}

\item{areaHarvested}{The value of area harvested.}
}
\description{
This is a function to compute yield based on production and area
harvested.
}

