% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sumWithNA}
\alias{sumWithNA}
\title{This is a wrapper for the normal sum function.}
\usage{
sumWithNA(x)
}
\arguments{
\item{x}{The vector to be summed.}
}
\description{
The function sums a vector with missing values, but when the whole
vector is missing it returns NA rather than zero.
}

